Recursion Maze Solver
This programs creates a maze randomly and uses recursion to find a path from the Start to the Destination
It contains 3 classes
MazeInput-Initialize and validate the user input for number of rows and column
MazeSolver-Create a maze randomly and find the path. If path is found, sum up the total points of the path.
LA2Main- The main class that creates the instance of MazeInput and MazeSolver.